Fix pages being rendered into the visible border radius during animations
When switching between Settings and Home Page view the animation would
show the pages being rendered into the border. This is
fixed by tightening the Clip element and fitting the pages
inside nicely.

Since the entire home page shifted to the right we can remove
the 16px x offset for the printer queue.