Add a basic settings page
This also includes a fix for the ComboBox to specify a minimum size.
Strictly speaking this isn't the minimum, as the text could be configured to elide,
so what we really want is a preferred size. But in the absence of that property
we must set the minimum-size to fix the column width in the settings grid layout.

Also fixed a few typos